@@ -1,12 +1,12 @@
-const log4js = require('log4js');
+import log4js from 'log4js';
 
-const loggerFactory = (category) => {
+const loggerFactory = (category: string | undefined) => {
   const logger = log4js.getLogger(category);
   logger.level = log4js.levels.INFO;
 
   return logger;
 };
 
-module.exports = {
+export {
   loggerFactory,
 };
